requirements.txt
pytest==7.4.3
pytest-xdist==3.5.0
selenium==4.15.2
allure-pytest==2.13.2
